legacy/aws_legacy/requirements.txt
boto3>=1.34.0
requests>=2.31.0
PyJWT[crypto]>=2.8.0
cryptography>=42.0.0
openpyxl>=3.1.0
python-multipart>=0.0.9
